web/requirements.txt
flask
opentelemetry-exporter-otlp
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask>=0.8b0
opentelemetry-instrumentation-requests>=0.8b0
requests

